Server test

+++ mortgage-application/Dto/ApplicantDto.cs
@@ -4,8 +4,8 @@ namespace mortgage_application.Dto
 {
     public class ApplicantDto
     {
-        [JsonPropertyName("principleAmount")] public long PrincipleAmount { get; set; }
-        [JsonPropertyName("annualRate")] public double AnnualRate { get; set; }
+        [JsonPropertyName("principleAmount")] public decimal PrincipleAmount { get; set; }
+        [JsonPropertyName("annualRate")] public decimal AnnualRate { get; set; }
         [JsonPropertyName("loanYears")] public int LoanYears { get; set; }
         [JsonPropertyName("startDate")] public DateTime? StartDate { get; set; }
     }
